working. will delete nan values in classes

+++ neiro2/neuron.h
@@ -7,9 +7,9 @@ public:
 	Neuron();
 	~Neuron();
 	int id;
-	float sum;
+	double sum;
 	std::vector<Link> inLinks; // -> (N)
-	float activation;
+	double activation;
 private:
 	
 };
